+++ core/app/models/spree/credit_card.rb
@@ -3,7 +3,6 @@ module Spree
     has_many :payments, as: :source
 
     before_save :set_last_digits
-    after_validation :set_card_type
 
     attr_accessor :number, :verification_value
 
@@ -14,36 +13,19 @@ module Spree
 
     scope :with_payment_profile, -> { where('gateway_customer_profile_id IS NOT NULL') }
 
+    def expiry=(expiry)
+      self[:month], self[:year] = expiry.split(" / ")
+      self[:year] = "20" + self[:year]
+    end
+
+    def number=(num)
+      @number = num.gsub(/[^0-9]/, '') rescue nil
+    end
+
     def set_last_digits
       number.to_s.gsub!(/\s/,'')
       verification_value.to_s.gsub!(/\s/,'')
       self.last_digits ||= number.to_s.length <= 4 ? number : number.to_s.slice(-4..-1)
-    end
-
-    # cheap hack to get to the type? method from deep within ActiveMerchant
-    # without stomping on potentially existing methods in CreditCard
-    class CardDetector
-      class << self
-        include ActiveMerchant::Billing::CreditCardMethods::ClassMethods
-      end
-    end
-
-    # Some payment gateways, such as USA EPay, only support an ActiveMerchant::Billing::CreditCard
-    # object, rather than an object *like* that. So we need to convert it.
-    def to_active_merchant
-      ActiveMerchant::Billing::CreditCard.new(
-        :number => number,
-        :month => month,
-        :year => year,
-        :verification_value => verification_value,
-        :first_name => first_name,
-        :last_name => last_name
-        )
-    end
-
-    # sets self.cc_type while we still have the card number
-    def set_card_type
-      self.cc_type ||= CardDetector.brand?(number)
     end
 
     def name?
@@ -94,9 +76,15 @@ module Spree
       gateway_customer_profile_id.present?
     end
 
-    def spree_cc_type
-      return 'visa' if Rails.env.development?
-      cc_type
+    def to_active_merchant
+      ActiveMerchant::Billing::CreditCard.new(
+        :number => number,
+        :month => month,
+        :year => year,
+        :verification_value => verification_value,
+        :first_name => first_name,
+        :last_name => last_name
+      )
     end
 
     private