+++ frontend/CHANGELOG.md
@@ -8,3 +8,7 @@
 *  Switch to new Google Analytics analytics.js SDK from ga.js SDK for custom dimensions & metrics.
 
    *Jeff Dutil*
+
+*  We now use [jQuery.payment](https://stripe.com/blog/jquery-payment) (from Stripe) to provide slightly better formatting on credit card number, expiry and CVV fields.
+
+   *Ryan Bigg*